test: only run the collaborative env when needed
In an effort to reduce the flakiness of the tests, we now only run tests with
multiple pages (multiple editors) when needed.
