+++ packages/editor/e2e-tests/__tests__/annotations-across-blocks.feature
@@ -1,7 +1,7 @@
 Feature: Annotations Across Blocks
 
   Background:
-    Given two editors
+    Given one editor
     And a global keymap
 
   Scenario: Adding annotation across blocks
