+++ packages/editor/e2e-tests/__tests__/annotations-overlapping.feature
@@ -1,7 +1,7 @@
 Feature: Overlapping Annotations
 
   Background:
-    Given two editors
+    Given one editor
     And a global keymap
 
   Scenario: Overlapping annotation
