+++ packages/editor/e2e-tests/__tests__/block-objects.feature
@@ -2,7 +2,7 @@
 Feature: Block Objects
 
   Background:
-    Given two editors
+    Given one editor
     And a global keymap
 
   Scenario: Pressing ArrowUp on a lonely image
