+++ packages/editor/e2e-tests/__tests__/decorators.feature
@@ -2,7 +2,7 @@
 Feature: Decorators
 
   Background:
-    Given two editors
+    Given one editor
     And a global keymap
 
   Scenario: Inserting text after a decorator
