+++ packages/editor/e2e-tests/__tests__/inline-objects.feature
@@ -2,7 +2,7 @@
 Feature: Inline Objects
 
   Background:
-    Given two editors
+    Given one editor
     And a global keymap
 
   # Currently fails
