+++ packages/editor/e2e-tests/__tests__/removing-blocks.feature
@@ -1,7 +1,7 @@
 Feature: Removing Blocks
 
   Background:
-    Given two editors
+    Given one editor
     And a global keymap
 
   Scenario: Pressing Delete in empty block with text below
