+++ packages/editor/e2e-tests/__tests__/removing-blocks.test.ts
@@ -1,4 +1,4 @@
-/** @jest-environment ./setup/collaborative.jest.env.ts */
+/** @jest-environment ./setup/jest.env.ts */
 
 import {Feature} from '@sanity/gherkin-driver/jest'
 import {parameterTypes, stepDefinitions} from './gherkin-step-definitions'
