+++ packages/editor/e2e-tests/__tests__/splitting-blocks.feature
@@ -1,7 +1,7 @@
 Feature: Splitting Blocks
 
   Background:
-    Given two editors
+    Given one editor
     And a global keymap
 
   Scenario: Splitting block at the beginning
